star_navi_backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
clearbit==0.1.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
Faker==4.0.2
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==7.1.1
psycopg2==2.8.5
pyhunter==1.7
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.25.8
